Flask/requirements.txt
flask
flask_cors
torch
numpy
librosa
tqdm
mediapipe
opencv_contrib_python
opencv_python
Pillow
torchvision
einops
timm
